% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defaults.R
\docType{data}
\name{acsdefaultendyearhere}
\alias{acsdefaultendyearhere}
\title{Default end year of 5-year ACS datasets this package targets}
\description{
Used as the default for \code{yr} arguments throughout the package. Update via
the script in \code{data-raw/datacreate_acsdefaultendyearhere.R} after each new
Census Bureau release.
}
\details{
See ACS release news:
\itemize{
\item 2020-2024 (released January 29, 2026): \url{https://www.census.gov/programs-surveys/acs/news/data-releases/2024/release-schedule.html}
\item 2019-2023 (released December 12, 2024): \url{https://www.census.gov/programs-surveys/acs/news/data-releases/2023/release.html}
}
}
