% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.sumlevel.R
\name{clean.sumlevel}
\alias{clean.sumlevel}
\title{Utility to Clean SUMLEVEL for get_acs_old}
\usage{
clean.sumlevel(sumlevel = "bg")
}
\arguments{
\item{sumlevel}{Character vector (1+ elements), optional, 'bg' by default. See details above.}
}
\value{
Returns 'both', 'tract', or 'bg' (or stops with error if cannot interpret sumlevel input)
}
\description{
Utility function used by \code{\link[=get_acs_old]{get_acs_old()}}.
}
\details{
Interprets as 'bg' any of these:    150, '150', 'blockgroup', 'block group', or 'bg' (or variants, ignoring case) \cr
Interprets as 'tract' any of these: 140, '140', or 'tract' (or variants, ignoring case) \cr
Interprets as 'both' any of these: 'both' or a vector that has both of the above terms (or variants, ignoring case).
}
\seealso{
\code{\link[=get_acs_old]{get_acs_old()}} which uses this
}
\keyword{internal}
