% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_est_moe.R
\name{format_est_moe}
\alias{format_est_moe}
\title{Reorder cols of estimates and MOE table}
\usage{
format_est_moe(my.list.of.tables)
}
\arguments{
\item{my.list.of.tables}{Required list of tables from earlier steps in \code{\link[=get.acs]{get.acs()}}}
}
\value{
List of tables like input but with columns sorted in a new order.
}
\description{
Start with a table that has all the estimates columns together, followed by all the MOEs columns,
and create a new column sort order so that estimates will be interspersed with (next to) their MOE values, as FactFinder format provides.
}
\seealso{
\code{\link[=get.acs]{get.acs()}} and \code{\link[analyze.stuff:intersperse]{analyze.stuff::intersperse()}}
}
