% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geofile.R
\name{geofile}
\alias{geofile}
\title{Get name(s) of GEO txt file(s) with geo information for ACS}
\usage{
geofile(mystates, end.year = acsdefaultendyearhere)
}
\arguments{
\item{mystates}{vector of character 2-letter State abbreviations specifying which are needed}

\item{end.year}{end.year of 5-year summary file such as '2021'}
}
\value{
Character vector of file names, example: "g20215md.txt"
Note this is only needed once per state, not once per seqfile. (It might even be available as a single US file?)
}
\description{
Get name of text file used by US Census Bureau with geographic information for American Community Survey.
That geo file can be used to join data file(s) to FIPS/GEOID/NAME/SUMLEVEL/CKEY.
}
\seealso{
\code{\link[=get_acs_old]{get_acs_old()}} and \code{\link[=download.geo]{download.geo()}} which uses this
}
\keyword{internal}
