% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoformat2010.R
\docType{data}
\name{geoformat2010}
\alias{geoformat2010}
\title{geographic information for 5-year summary file ACS dataset ending in given year}
\format{
A data.frame
\code{
'data.frame':	51 obs. of  3 variables:
  $ varname: chr  "FILEID" "STUSAB" "SUMLEVEL" "COMPONENT" ...
  $ size   : num  6 2 3 2 7 1 1 1 2 2 ...
  $ start  : num  1 7 9 12 14 21 22 23 24 26 ...
  }
}
\source{
\url{http://www2.census.gov/programs-surveys/acs}
}
\description{
This data set has the format used by geographic identifier files in the American Community Survey (ACS) 5-year summary file.
}
\seealso{
\code{\link{get.acs}}
}
\keyword{datasets}
