% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoformat2019.R
\docType{data}
\name{geoformat2019}
\alias{geoformat2019}
\title{geographic information for ACS dataset}
\format{
A data.frame
\preformatted{
 'data.frame':	53 obs. of  5 variables:
 $ varname    : chr  "FILEID" "STUSAB" "SUMLEVEL" "COMPONENT" ...
 $ description: chr  "Always equal to ACS Summary File identification" "State Postal Abbreviation" "Summary Level" "Geographic Component" ...
 $ size       : num  6 2 3 2 7 1 1 1 2 2 ...
 $ start      : num  1 7 9 12 14 21 22 23 24 26 ...
 $ type       : chr  "Record" "Record" "Record" "Record" ...
  }
}
\source{
Table found in given year dataset, info at
\url{https://www.census.gov/programs-surveys/acs/library/handbooks/geography.html}
\url{https://www.census.gov/programs-surveys/acs/data/data-via-ftp.html}
\url{https://www2.census.gov/programs-surveys/acs/summary_file/2020/data/5_year_entire_sf/2020_ACS_Geography_Files.zip}
}
\description{
This data set has the format used by geographic identifier files
in the American Community Survey (ACS) 5-year summary file.
The data and documentation for the 5 years ending in year X
is typically available by December of the year X+1, so 2015-2019 was available by Dec 2020.
}
\seealso{
\code{\link[=get.acs]{get.acs()}}
}
\keyword{datasets}
