% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.acs.all.R
\name{get.acs.all}
\alias{get.acs.all}
\title{get ACS summaryfile specified tables for entire USA, all blockgroups or all tracts}
\usage{
get.acs.all(
  tables = "B01001",
  end.year = 2020,
  dataset = "5",
  sumlevel = 150,
  output.path = file.path("~", "acsoutput")
)
}
\arguments{
\item{tables}{'ejscreen' or else one or more table codes like 'b01001' or c("B01001", "B03002")}

\item{end.year}{2020 or another available year}

\item{dataset}{default is 5, for 5-year summary file data like 2016-2020 ACS}

\item{sumlevel}{default is 150 for blockgroups. 140 is tracts.}

\item{output.path}{folder / path where to save the downloaded files}
}
\value{
list of data.frames, one per table requested
}
\description{
Uses Census Bureau new format files that offer the full USA in 1 file per table,
via FTP or https.
Does not get Puerto Rico (PR) as coded.
Note API provides better list of variable names, in order as found on summary files, than documentation does.
}
\details{
For information on this new Summary File format visit:

https://www.census.gov/programs-surveys/acs/data/summary-file.html
}
