% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.bg.R
\name{get.bg}
\alias{get.bg}
\title{Get just the Census block group part of existing ACS data}
\usage{
get.bg(merged.tables.mine)
}
\arguments{
\item{merged.tables.mine}{Required set of tables in format used by \code{\link[=get_acs_old]{get_acs_old()}}}
}
\value{
subset of the inputs, same format
}
\description{
Helper function to return just the block group resolution part of a dataset in \code{\link[=get_acs_old]{get_acs_old()}}
}
\seealso{
\code{\link[=get_acs_old]{get_acs_old()}}, \code{\link[=get.tracts]{get.tracts()}}
}
\keyword{internal}
