% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.lookup.file.name.R
\name{get.lookup.file.name}
\alias{get.lookup.file.name}
\title{Get name of Census file with Sequence and Table Numbers for ACS 5-year summary file data}
\usage{
get.lookup.file.name(end.year = acsdefaultendyearhere)
}
\arguments{
\item{end.year}{Not yet implemented, but will be optional end year for 5-year summary file, as character}
}
\value{
Returns character element that is name of file such as "Sequence_Number_and_Table_Number_Lookup.txt"
}
\description{
Returns name of text file provided by US Census Bureau,
such as Sequence_Number_and_Table_Number_Lookup.txt, which provides the
sequence numbers (file numbers) and table numbers for data in the American Community Survey (ACS) 5-year summary file.
}
\seealso{
\code{\link[=get_acs_old]{get_acs_old()}}, \code{\link[=get.lookup.acs]{get.lookup.acs()}}, \code{\link[=get.url.prefix.lookup.table]{get.url.prefix.lookup.table()}}. Also see \code{data(lookup.acs)}.
}
\keyword{internal}
