% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.url.prefix.lookup.table.R
\name{get.url.prefix.lookup.table}
\alias{get.url.prefix.lookup.table}
\title{Get URL (without filename) for ACS 5-year summary file data}
\usage{
get.url.prefix.lookup.table(end.year)
}
\arguments{
\item{end.year}{Optional end year for 5-year summary file, as character, like '2022'}
}
\value{
Returns character element that is first part of URL such as
"https://www2.census.gov/programs-surveys/acs/summary_file/2022/table-based-SF/documentation/"
}
\description{
Returns first part of URL(s) of folders (on Census Bureau FTP site) with zip file(s) based on end year.
}
\seealso{
\code{\link[=get_acs_old]{get_acs_old()}}, \code{\link[=get.lookup.acs]{get.lookup.acs()}}, \code{\link[=get.lookup.file.name]{get.lookup.file.name()}}
}
\keyword{internal}
