% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.zipfile.prefix.R
\name{get.zipfile.prefix}
\alias{get.zipfile.prefix}
\title{Get first part of ACS zip file name.}
\usage{
get.zipfile.prefix(end.year = acsdefaultendyearhere)
}
\arguments{
\item{end.year}{Optional character, specifying last year of 5-year summary file data.}
}
\description{
Get the first part of the zipfile name for the ACS 5-year summary file datafiles on the US Census Bureau FTP site.
}
\seealso{
\code{\link[=get_acs_old]{get_acs_old()}}, \code{\link[=get.datafile.prefix]{get.datafile.prefix()}}, \code{\link[=datafile]{datafile()}}, \code{\link[=geofile]{geofile()}}, \code{\link[=get.zipfile.prefix]{get.zipfile.prefix()}}
}
\keyword{internal}
