% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{get_acs_old}
\alias{get_acs_old}
\title{Defunct: download ACS data in the pre-2022 sequence-file format}
\usage{
get_acs_old(...)
}
\arguments{
\item{...}{ignored}
}
\value{
This function always stops with an error.
}
\description{
\code{get_acs_old()} and the entire sequence-file code path were removed in
ACSdownload 3.0.0. Use \code{\link[=get_acs_new]{get_acs_new()}}, which reads the Census Bureau's
table-based summary file format (available for end year 2022 and later).
}
\details{
If you specifically need the old sequence-file format for an end year of
2021 or earlier, install the final 2.x release:
\code{remotes::install_github("ejanalysis/ACSdownload@v2.4.0-pre-refactor")}.
}
\keyword{internal}
