% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getstatesviafilenames.R
\name{getstatesviafilenames}
\alias{getstatesviafilenames}
\title{Infer US States based on ACS 5-yr filenames in folder}
\usage{
getstatesviafilenames(folder = getwd())
}
\arguments{
\item{folder}{Default is current working directory.}
}
\value{
Returns a vector of unique upper case US State abbreviations
}
\description{
Helper function to look for unzipped csv files of estimates for American Community Survey (ACS)
5-year summary file data obtained from US Census FTP site, based on pattern matching, and infer State abbreviations
based on those filenames.
}
\seealso{
\code{\link[=read.concat.states]{read.concat.states()}} which uses this
}
\keyword{internal}
