% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gettablesviaseqnums.R
\name{gettablesviaseqnums}
\alias{gettablesviaseqnums}
\title{Look up ACS table IDs of tables in given sequence number files}
\usage{
gettablesviaseqnums(x, end.year = acsdefaultendyearhere)
}
\arguments{
\item{x}{Required character vector of table IDs such as "B01001"}

\item{end.year}{Optional character variable providing the end year of the 5-year ACS survey,
to ensure the proper sequence file to table ID matching is used.}
}
\value{
Returns a character vector of table IDs such as "B01001"
}
\description{
Helper function to look for which American Community Survey (ACS) tables are in given sequence files
as obtained from the US Census FTP site.
}
\seealso{
\code{\link[=read.concat.states]{read.concat.states()}} which uses this, and \code{\link[=get.lookup.acs]{get.lookup.acs()}} which is used by this
}
\keyword{internal}
