% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_files_ftp.R
\name{list_files_ftp}
\alias{list_files_ftp}
\title{list_files_ftp
Attempt to get list of filenames on an FTP site}
\usage{
list_files_ftp(
  url,
  credentials = "",
  sleep = NA,
  sort = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{url}{URL}

\item{credentials}{empty}

\item{sleep}{NA}

\item{sort}{FALSE}

\item{verbose}{FALSE}
}
\value{
file names
}
\description{
list_files_ftp
Attempt to get list of filenames on an FTP site
}
