% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_files_ftp.R
\name{list_files_ftp_worker}
\alias{list_files_ftp_worker}
\title{list_files_ftp_worker
Attempt to get list of filenames on an FTP site. See list_files_ftp()
Uses RCurl::getURL(url, dirlistonly = TRUE)}
\usage{
list_files_ftp_worker(url, credentials, sleep, verbose)
}
\arguments{
\item{url}{URL}

\item{credentials}{?}

\item{sleep}{?}

\item{verbose}{?}
}
\value{
list of file names
}
\description{
list_files_ftp_worker
Attempt to get list of filenames on an FTP site. See list_files_ftp()
Uses RCurl::getURL(url, dirlistonly = TRUE)
}
