% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookup.acs2019.R
\docType{data}
\name{lookup.acs2019}
\alias{lookup.acs2019}
\title{ACS_5yr_Seq_Table_Number_Lookup.txt for ACS dataset}
\description{
ACS_5yr_Seq_Table_Number_Lookup.txt for ACS dataset
}
\seealso{
\link{lookup.acs} and  \code{\link[acs:acs.lookup]{acs::acs.lookup()}} \code{\link[tidycensus:load_variables]{tidycensus::load_variables()}} \code{\link[=get.lookup.acs]{get.lookup.acs()}} \code{\link[=get.acs]{get.acs()}}
}
