% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookup.acs2021.R
\docType{data}
\name{lookup.acs2021}
\alias{lookup.acs2021}
\title{ACS_5yr_Seq_Table_Number_Lookup.txt for ACS dataset}
\description{
ACS_5yr_Seq_Table_Number_Lookup.txt for ACS dataset
}
\details{
created via download of \link{https://www2.census.gov/programs-surveys/acs/summary_file/2021/sequence-based-SF/documentation/user_tools/ACS_5yr_Seq_Table_Number_Lookup.txt}

lookup.acs2021 <- read.csv("ACSdownload/inst/2021ACS_5yr_Seq_Table_Number_Lookup.txt")
usethis::use_data(lookup.acs2021)
lookup.acs <- lookup.acs2021
usethis::use_data(lookup.acs, overwrite = TRUE)
}
\seealso{
\link{lookup.acs} and  \code{\link[acs:acs.lookup]{acs::acs.lookup()}} \code{\link[tidycensus:load_variables]{tidycensus::load_variables()}} \code{\link[=get.lookup.acs]{get.lookup.acs()}} \code{\link[=get.acs]{get.acs()}}
}
