% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookup.acs2022.R
\docType{data}
\name{lookup.acs2022}
\alias{lookup.acs2022}
\title{ACS_5yr_Seq_Table_Number_Lookup.txt for ACS dataset}
\description{
ACS_5yr_Seq_Table_Number_Lookup.txt for ACS dataset
}
\details{
WORK IN PROGRESS BECAUSE  Formats available changed!

ACS2018-2022 does not offer sequence file style downloads.
Now available only by table, requiring a different approach to download
data.
The format of the lookup.acs2022 file is new so the older code
for prior years will not work.

\if{html}{\out{<div class="sourceCode">}}\preformatted{lookup.acs2022 <- read.delim("ACSdownload/inst/ACS20225YR_Table_Shells.txt", sep = "|")

usethis::use_data(lookup.acs2022)
}\if{html}{\out{</div>}}

See \link{https://www2.census.gov/programs-surveys/acs/summary_file/2022/table-based-SF/documentation/}
}
\seealso{
\link{lookup.acs} and  \code{\link[acs:acs.lookup]{acs::acs.lookup()}} \code{\link[tidycensus:load_variables]{tidycensus::load_variables()}} \code{\link[=get.lookup.acs]{get.lookup.acs()}} \code{\link[=get.acs]{get.acs()}}
}
