% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge.tables.R
\name{merge.tables}
\alias{merge.tables}
\title{Concatenate several ACS data tables into one big table}
\usage{
\method{merge}{tables}(my.list.of.tables)
}
\arguments{
\item{my.list.of.tables}{Required list of data tables from prior steps in \code{\link{get.acs}}}
}
\value{
Returns one big data.frame with all columns of all input tables
}
\description{
Concatenate several ACS data tables into one big table
}
\seealso{
\code{\link{get.acs}}
}

