% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/read.nhgis.R
\name{read.nhgis}
\alias{read.nhgis}
\title{Read NHGIS.org ACS Data Files and Codebooks}
\usage{
read.nhgis(datafile, codebookfile = gsub("\\\\.csv", "_codebook.txt",
  datafile), folder = getwd())
}
\arguments{
\item{datafile}{Names of files}

\item{codebookfile}{Optional name(s) of codebook files. Default is to infer from datafile}

\item{folder}{Optional path where files are found. Default is getwd()}
}
\value{
Returns a named list: data, contextfields, fields, tables, geolevel, years, dataset
}
\description{
Helper function used by nhgisimport to read downloaded and unzipped csv and txt files obtained from NHGIS.org,
  with US Census Bureau data from the American Community Survey (ACS).
}
\seealso{
\code{\link{nhgisimport}} which uses this, \code{\link{get.acs}}, \code{\link{get.datafile.prefix}}, \code{\link{datafile}}, \code{\link{geofile}}, \code{\link{get.zipfile.prefix}}
}

