% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unzip.datafiles.R
\name{unzip.datafiles}
\alias{unzip.datafiles}
\title{Unzip ACS datafile per state, downloading missing ones first}
\usage{
unzip.datafiles(
  tables,
  mystates,
  folder = getwd(),
  end.year = acsdefaultendyearhere,
  testing = FALSE,
  attempts = 5,
  silent = FALSE
)
}
\arguments{
\item{tables}{Character vector of table numbers needed such as 'B01001'}

\item{mystates}{Character vector of 2-character state abbreviations. Default is all states.}

\item{folder}{Default is current working directory.}

\item{end.year}{optional -- specifies last year of 5-year summary file.}

\item{testing}{Default is FALSE. If TRUE, prints more info.}

\item{attempts}{Default is 5, specifies how many tries (maximum) for unzipping before trying to redownload and then give up.}

\item{silent}{Default is FALSE. Whether to send progress info to standard output.}
}
\value{
Side effect is unzipping file on disk (unless testing=TRUE)
}
\description{
Unzip ACS datafile for each specified US State, extracting specified table(s), downloading missing zip files first.
}
\seealso{
\code{\link[=get_acs_old]{get_acs_old()}}
}
\keyword{internal}
