% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/url_acs_table.R
\name{url_acs_table}
\alias{url_acs_table}
\title{Build data.census.gov URLs for one or more ACS 5-year tables}
\usage{
url_acs_table(
  tables = ejscreen_acs_tables,
  fips = NULL,
  yr = acsdefaultendyearhere,
  fiveorone = 5
)
}
\arguments{
\item{tables}{one or more table codes like "B01001" or c("B01001", "C17002")}

\item{fips}{optional vector of fips codes such as "34023001419"; NULL or
empty string means "just the general info URL for each table"}

\item{yr}{end year of the 5-year ACS summary file}

\item{fiveorone}{1 or 5}
}
\value{
character vector of URLs, one per (table x fips) pair
}
\description{
Produces URLs to either general information about a table or a
table-for-fips view.
}
\examples{
  url_acs_table()
  # browseURL(url_acs_table(tables = "C17002", yr = 2024))
}
\seealso{
\code{EJAM::url_acs_table_info()} for a richer EJSCREEN-aware variant
}
