% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/which.seqfiles.R
\name{which.seqfiles}
\alias{which.seqfiles}
\title{Find Which Sequence Files Contain Given ACS Table(s)}
\usage{
which.seqfiles(tables, lookup.acs, end.year = acsdefaultendyearhere)
}
\arguments{
\item{tables}{character vector, required. Defines which ACS table(s) to check, such as 'B01001'}

\item{lookup.acs}{data.frame, optional (if not provided then it is downloaded from Census).
Specifies what variables are in which tables and which tables are in which sequence files on the FTP site.}

\item{end.year}{Character element, optional. Defines end year for 5-year dataset.
Valid years are limited.
Ignored if lookup.acs is specified, however. If they imply different years, the function stops with an error message.}
}
\value{
Returns a vector of one or more numbers stored as characters, each defining one sequence file, such as "0001".
}
\description{
The US Census Bureau provides 5-year summary file data from the American Community Survey in sequence files on their FTP site.
This function reports which sequence files contain the specified tables. Used by \code{\link[=get_acs_old]{get_acs_old()}}
}
\seealso{
\code{\link[=get_acs_old]{get_acs_old()}} and  from the \pkg{acs} package, which does something related but is more flexible & robust. Also see \code{\link[=get_acs_old]{get_acs_old()}} which uses this.
}
\keyword{internal}
