% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acsdefaultendyearhere.R
\docType{data}
\name{acsdefaultendyearhere}
\alias{acsdefaultendyearhere}
\title{The default end.year of 5yr ACS datasets this package uses in functions}
\description{
This is to make it easier to update the package over time.
It should be updated as needed, along with acsfirstyearavailablehere
}
