% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ACSdownload-package.R
\docType{package}
\name{ACSdownload}
\alias{ACSdownload}
\alias{ACSdownload-package}
\title{Obtain American Community Survey Summary File Data Tables from Census Bureau FTP Site}
\description{
THIS PACKAGE MAY BE MOSTLY OBSOLETE NOW THAT SOME CRAN PACKAGES OFFER MOST OF THESE TOOLS.
}
\details{
\itemize{
\item ACSdownload is an R package that helps you download and
parse huge raw data from the
Census Bureau American Community Survey 5-year Summary Files,
providing demographic data at the block and tract levels of resolution.
You can obtain data from the entire USA all at once using this package,
for one or more tables. Some key functions are \code{\link[=get.acs]{get.acs()}}, \code{\link[=set.needed]{set.needed()}}, and \code{\link[=nhgis]{nhgis()}}.
Also see data(\link{lookup.acs2022}).
\cr\cr
\item The (tidycensus package)\link{https://walker-data.com/tidycensus/index.html}
makes it easy to download some ACS or decennial census or other data,
(once you get a Census Bureau API key to request the data), but
downloading via API is slow/awkward if you want all blockgroups nationwide.
\cr\cr
\item Also, there was a package called totalcensus that tried to
offer tools to download ACS5yr tables,
but as of 8/2025 that package had not been updated since 12/2023,
and the late 2023 version did not work out of the box.
The \href{https://cran.r-project.org/web/packages/totalcensus/index.html}{CRAN version of totalcensus}
was older than the \href{https://github.com/GL-Li/totalcensus}{github version of totalcensus}
\code{devtools::install_github("GL-Li/totalcensus")}
\cr\cr
\item The Census Bureau makes it easy to obtain data from one
state at a time, not every block group in the US.
There are over 240,000 block groups and over 85,000 tracts in the U.S. \cr
\cr
HOWEVER, THE DATA FORMAT HAS CHANGED FOR SUMMARY FILE ACS DATA: \cr
\link{https://www.census.gov/programs-surveys/acs/data/summary-file/updates-to-acs-summary-file.Overview.html} \cr
\cr
Several options for obtaining Census ACS data are now listed here: \cr
\link{http://www.census.gov/programs-surveys/acs/data.html} \cr
Limits on downloads via American Fact Finder, not all US tracts at once
are noted here: \link{https://ask.census.gov/faq.php?id=5000&faqId=1653} \cr
Other data sources that may be relevant include Census geodatabases at \cr
\link{http://www.census.gov/geo/maps-data/data/tiger-data.html} and data at \cr
\link{http://www.census.gov/geo/maps-data/data/gazetteer.html}.
}
}
\references{
\link{https://ejanalysis.org} \cr
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ejanalysis/ACSdownload}
  \item \url{https://ejanalysis.github.io/ACSdownload/}
  \item \url{https://ejanalysis.org}
}

}
\author{
info@ejanalysis.com
}
