% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ACSdownload-package.R
\docType{package}
\name{ACSdownload}
\alias{ACSdownload}
\alias{ACSdownload-package}
\title{Obtain American Community Survey Summary File Data Tables from Census Bureau FTP Site}
\description{
THIS PACKAGE MAY BE MOSTLY OBSOLETE NOW THAT SOME CRAN PACKAGES OFFER
MOST OF THESE TOOLS.
}
\details{
There is a package called totalcensus that tries to
offer tools to download ACS5yr tables, but the CRAN and dev github versions
have bugs as of 8/2023, and did not work out of the box.
totalcensus \link{https://cran.r-project.org/web/packages/totalcensus/index.html}
see   devtools::install_github("GL-Li/totalcensus")

ACSdownload is an R package that helps you download and
parse huge raw data from the
Census Bureau American Community Survey 5-year Summary Files,
providing demographic data at the block and tract levels of resolution.

You can obtain data from the entire USA all at once using this package,
for one or more tables.

Typically the Census Bureau makes it easy to obtain data from one
state at a time, not every block group in the US. There are roughly 220,000
block groups in the US, and around 74,000 tracts

Some options for obtaining Census ACS data are listed here:
\link{http://www.census.gov/programs-surveys/acs/data.html}
Other data sources that may be relevant include Census geodatabases
at \link{http://www.census.gov/geo/maps-data/data/tiger-data.html} and data at
\link{http://www.census.gov/geo/maps-data/data/gazetteer.html}
For any imported/suggested packages not on CRAN, see \link{https://ejanalysis.github.io}

Key function is \code{\link[=get.acs]{get.acs()}}, and
also see \code{\link[=set.needed]{set.needed()}}, \code{\link[=nhgis]{nhgis()}},
and also see data(\link{lookup.acs2020})
HOWEVER, DATA FORMAT IS CHANGING FOR SUMMARY FILE ACS DATA:
\link{https://www.census.gov/programs-surveys/acs/data/summary-file/}
updates-to-acs-summary-file.Overview.html

Typically the Census Bureau makes it easy to obtain data from one state at a time,
not every block group in the US. There are roughly 220,000 block groups in the US,
and around 74,000 tracts.
The key function in this package is \code{\link[=get.acs]{get.acs()}} \cr\cr
For ACS documentation, see \link{http://www.census.gov/programs-surveys/acs.html} \cr

Several options for obtaining Census ACS data are now listed here: \cr

\link{http://www.census.gov/programs-surveys/acs/data.html} \cr
Limits on downloads via American Fact Finder, not all US tracts at once
are noted here: \link{https://ask.census.gov/faq.php?id=5000&faqId=1653} \cr

Other data sources that may be relevant include Census geodatabases at \cr
\link{http://www.census.gov/geo/maps-data/data/tiger-data.html} and data at \cr
\link{http://www.census.gov/geo/maps-data/data/gazetteer.html}.

\if{html}{\out{<div class="sourceCode">}}\preformatted{Also see the help for [get.acs()] \\cr\\cr
}\if{html}{\out{</div>}}
}
\references{
\link{http://ejanalysis.github.io} \cr
\link{http://www.ejanalysis.com} \cr
}
\seealso{
\pkg{[tidycensus][tidycensus::tidycensus]} to use a key to request ACS
or Decennial Census data, but is slow if you want all blockgroups in a state or nationwide.
\pkg{proxistat} package for block group points as lat lon, or
\pkg{acs} package \link{http://cran.r-project.org/web/packages/acs/index.html}
which lets one obtain more limited amounts of ACS data
but provides better tools for working with the data once obtained.
}
\author{
info@ejanalysis.com
}
