% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ACSdownload-package.R
\docType{package}
\name{ACSdownload}
\alias{ACSdownload}
\alias{ACSdownload-package}
\title{Obtain American Community Survey Summary File Data Tables from Census Bureau FTP Site}
\description{
This R package helps you download and parse raw data from the
  Census Bureau American Community Survey 5-year Summary Files,
  providing demographic data at the block and tract levels of resolution.
  You can obtain data from the entire USA all at once using this package, for one or more tables.
  Typically the Census Bureau makes it easy to obtain data from one state at a time,
  not every block group in the US. There are roughly 220,000 block groups in the US,
  and around 74,000 tracts.
  Several options for obtaining Census ACS data are now listed here:
  \url{http://www.census.gov/programs-surveys/acs/data.html} \cr
  Limits on downloads via American Fact Finder (not all US tracts at once) are noted here: \url{https://ask.census.gov/faq.php?id=5000&faqId=1653} \cr
  Other data sources that may be relevant include Census geodatabases at
  \url{http://www.census.gov/geo/maps-data/data/tiger-data.html} and data at
  \url{http://www.census.gov/geo/maps-data/data/gazetteer.html}. Also see the help for \code{\link{get.acs}} \cr\cr
  The key function in this package is \code{\link{get.acs}} \cr\cr
  For ACS documentation, see \url{http://www.census.gov/programs-surveys/acs.html}
}
\author{
info@ejanalysis.com <info@ejanalysis.com>
}
\references{
\url{http://ejanalysis.github.io} \cr
 \url{http://www.ejanalysis.com} \cr
}
\seealso{
\pkg{acs} package (\url{http://cran.r-project.org/web/packages/acs/index.html})
  which lets one obtain more limited amounts of ACS data
  but provides better tools for working with the data once obtained.
}

