% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acsfirstyearavailablehere.R
\docType{data}
\name{acsfirstyearavailablehere}
\alias{acsfirstyearavailablehere}
\title{The earliest end.year of 5yr ACS datasets this package handles}
\description{
This is to make it easier to update the package over time.
It should be updated as needed, along with acsdefaultendyearhere
}
