% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.mystates.R
\name{clean.mystates}
\alias{clean.mystates}
\title{Utility to Clean Names of States for get_acs_old}
\usage{
clean.mystates(mystates = "all", testing = FALSE)
}
\arguments{
\item{mystates}{Character vector, optional.
Defines which states, using 2-character abbreviations (case-insensitive), or 'all' for all available.

Default is 'all' which is 50 states plus DC, BUT NO LONGER PR (& not AS, GU, MP, UM, VI, US).
If mystates is not specified or is 'all' then this returns the default.
If mystates is specified, returns those defaults that match any of mystates.
If mystates includes DC, PR, or US, those are returned,
but "AS" "GU" "MP" "VI" are removed,
and "UM" is removed.
One element of the vector can be 50, 51, or 52, where 50 represents
the 50 states, 51 = 50 plus DC, or 52 = the 51 plus PR.
In other words, the default could be written as c(50,'DC','PR')
or as c(51,'PR') or just 52.
Redundant entries are dropped, e.g., c(51,'DC') becomes 51.}

\item{testing}{Logical value, optional, FALSE by default. LIMITS STATES TO DC AND DE if TRUE.}
}
\value{
Returns character vector of 2-character State abbreviations, lower case.
}
\description{
Utility function used by \code{\link[=get_acs_old]{get_acs_old()}} and \code{\link[=download.datafiles]{download.datafiles()}}
}
\details{
Not in FTP ACS Summary files and not returned by this function's default: \cr
#53         60 AS              American Samoa               \if{html}{\out{<NA>}}\cr
#54         66 GU                        Guam               \if{html}{\out{<NA>}}\cr
#55         69 MP    Northern Mariana Islands               \if{html}{\out{<NA>}}\cr
#58         78 VI         U.S. Virgin Islands               \if{html}{\out{<NA>}}\cr
#57         74 UM U.S. Minor Outlying Islands               \if{html}{\out{<NA>}}\cr

Note: The function stops if any of mystates is not found in the full list that includes these,
but it returns only those excluding these five above, so an invalid state like 'ZQ' causes an error but an invalid state like 'VI' is silently removed without any error.

Puerto Rico (PR, FIPS 72), the District of Columbia (DC, 11), and the 50 States are in ACS 2016-2020 and Census 2020.

For info on Island Areas, see
\url{https://www.census.gov/programs-surveys/decennial-census/technical-documentation/island-areas-censuses.html}
The 2020 Island Areas Censuses (IAC) include data for American Samoa (AS, FIPS 60), Guam (GU, 66),
the Commonwealth of the Northern Mariana Islands (MP, 69), and the U.S. Virgin Islands (VI. 78),
but not U.S. Minor Outlying Islands (UM, 74).
The 2020 Census counted people living in the U.S. Island Areas using a long-form questionnaire. Other
surveys, such as the American Community Survey (ACS), are not conducted in the Island Areas. Therefore,
the Census Bureau used a long-form questionnaire to meet the Island Areas’ data needs for demographic,
social, economic, and housing unit information. This long-form questionnaire was similar to the ACS
questionnaire used in the 50 states, the District of Columbia, and Puerto Rico.
The 2020 IAC Demographic Profile Summary File provides access to all of the data.
It provides data down to the place, county subdivision, and estate (USVI only) level - not block groups.
The 2020 IAC Demographic Profile Summary File data are available through
the Census Bureau’s data exploration platform, data.census.gov.
The 2020 IAC Demographic Profile Summary File is located on the U.S. Census Bureau’s file transfer protocol
(FTP) server at \url{https://www2.census.gov/programs-surveys/decennial/2020/data/island-areas/}

You can start at the \href{https://www.census.gov/programs-surveys/decennial-census/decade/2020/planning-management/release/island-area-data-products-faqs.html}{2020 Island Areas Censuses Data Products FAQ}
}
\seealso{
\code{\link[=get_acs_old]{get_acs_old()}} and \code{\link[=download.datafiles]{download.datafiles()}} which use this
}
