% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.mystates.R
\name{clean.mystates}
\alias{clean.mystates}
\title{Utility to Clean Names of States for get.acs}
\usage{
clean.mystates(mystates = "all", testing = FALSE)
}
\arguments{
\item{mystates}{Character vector, optional. 
Defines which states, using 2-character abbreviations (case-insensitive), or 'all' for all available.

<<<<<<< HEAD
Default is 'all' which is 50 states plus DC plus PR (but not AS, GU, MP, UM, VI, US).
If mystates is not specified or is 'all' then this returns the default.
If mystates is specified, returns those defaults that match any of mystates.
If mystates includes DC, PR, or US, those are returned, but "AS" "GU" "MP" "UM" "VI" are removed.
One element of the vector can be 50, 51, or 52, which represents the 50 states, those 50 plus DC, or the 51 plus PR. 
=======
Default is 'all' which is 50 states plus DC, BUT NO LONGER PR (& not AS, GU, MP, UM, VI, US).
If mystates is not specified or is 'all' then this returns the default.
If mystates is specified, returns those defaults that match any of mystates.
If mystates includes DC, PR, or US, those are returned, but "AS" "GU" "MP" "UM" "VI" are removed.
One element of the vector can be 50, 51, or 52, where 50 represents the 50 states, 51 = 50 plus DC, or 52 = the 51 plus PR.
>>>>>>> 432b60ac63f2153b90bc23f75c01f62c09fd2160
In other words, the default could be written as c(50,'DC','PR') or as c(51,'PR') or just 52.
Redundant entries are dropped, e.g., c(51,'DC') becomes 51.}

\item{testing}{Logical value, optional, FALSE by default. LIMITS STATES TO DC AND DE if TRUE.}
}
\value{
Returns character vector of 2-character State abbreviations, lower case.
}
\description{
Utility function used by \code{\link{get.acs}} and \code{\link{download.datafiles}}
}
\details{
Not in FTP ACS Summary files and not returned by this function's default: \cr
 #53         60 AS              American Samoa               <NA>\cr
 #54         66 GU                        Guam               <NA>\cr
 #55         69 MP    Northern Mariana Islands               <NA>\cr
 #57         74 UM U.S. Minor Outlying Islands               <NA>\cr
 #58         78 VI         U.S. Virgin Islands               <NA>\cr
 Note: The function stops if any of mystates is not found in the full list that includes these,
 but it returns only those excluding these five above, so an invalid state like 'ZQ' causes an error but an invalid state like 'VI' is silently removed without any error.
}
\seealso{
\code{\link{get.acs}} and \code{\link{download.datafiles}} which use this, and \code{\link[ejanalysis]{get.state.info}} (from \pkg{ejanalysis} package) based on \code{\link[proxistat]{lookup.states}} or \code{data(lookup.states, package=proxistat)} using \pkg{proxistat} package
}
