% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/datafile.R
\name{datafile}
\alias{datafile}
\title{Get name(s) of data file(s) for ACS 5-year summary file data}
\usage{
datafile(state.abbrev, seqfilenum, end.year = "2012")
}
\arguments{
\item{state.abbrev}{Required vector of 2-character state abbreviation(s)}

\item{seqfilenum}{Required single sequence file number used by ACS 5-year summary file}

\item{end.year}{Optional end year for 5-year summary file, as character, defaulting to "2012"}
}
\value{
Returns character element that is name of data file such as e20105de0017000 or m20105de0017000
}
\description{
Returns name(s) of data file(s) based on state(s), a sequence file number, and end year.
}
\seealso{
\code{\link{get.acs}}
}

