% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datafile.R
\name{datafile}
\alias{datafile}
\title{Get name(s) of data file(s) for ACS 5-year summary file data}
\usage{
datafile(state.abbrev, seqfilenum, end.year = "2019")
}
\arguments{
\item{state.abbrev}{Required vector of one or more 2-character state abbreviations like "DC"}

\item{seqfilenum}{Required sequence file number(s) used by ACS 5-year summary file (can be a single value like "0022" or a vector)}

\item{end.year}{Optional end year for 5-year summary file, as character or number}
}
\value{
Returns character element that is name of data file such as e20105de0017000 or m20105de0017000
}
\description{
Returns name(s) of data file(s) based on state(s), a sequence file number, and end year.
}
\seealso{
\code{\link{get.acs}}
}
