% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.datafiles.R
\name{download.datafiles}
\alias{download.datafiles}
\title{Download American Community Survey 5-yr data files}
\usage{
download.datafiles(tables, end.year = "2012", mystates = 52,
  folder = getwd(), testing = FALSE, attempts = 5, silent = FALSE)
}
\arguments{
\item{tables}{Required character vector of table numbers, such as c("B01001", "B03002")}

\item{end.year}{Character element, optional, "2012" by default. Defines last of five years of summary file dataset; default is 2008-2012.}

\item{mystates}{Character vector, now optional - Default is 50 states + DC + PR here, but otherwise relies on \code{\link{clean.mystates}}}

\item{folder}{Default is getwd()}

\item{testing}{Default to FALSE. If TRUE, provides info on progress of download.}

\item{attempts}{Default is 5, specifies how many tries (maximum) for unzipping before trying to redownload and then give up.}

\item{silent}{Optional, default is FALSE. Whether progress info should be sent to standard output (like the screen).}
}
\value{
Effect is to download and save locally a number of data files.
}
\description{
Attempts to download data files (estimates and margins of error) for specified states and tables,
  from the US Census Bureau's FTP site for American Community Survey (ACS) 5-year summary file data.
}
\seealso{
\code{\link{get.distances}} which allows you to get distances between all points.
}

