% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.lookup.acs.R
\name{download.lookup.acs}
\alias{download.lookup.acs}
\title{Download File with Information about ACS 5-Year Summary File Tables}
\usage{
download.lookup.acs(
  end.year = acsdefaultendyearhere,
  folder = NULL,
  silent = FALSE
)
}
\arguments{
\item{end.year}{Character, optional, like '2020', which specifies the 2016-2020 dataset.
Defines which 5-year summary file to use, based on end-year.
Can be acsfirstyearavailablehere or later. Data for end.year='2019' were released in December 2020, for example.
The 2017-2021 American Community Survey 5-year estimates are scheduled to be released on Thursday, December 8, 2022.}

\item{folder}{Optional path to where to download file to, defaults to temp folder.}

\item{silent}{Optional, default is FALSE. Whether to send progress info to standard output (like the screen)}
}
\value{
By default, returns a data.frame with these fields:
\itemize{
\item $ Table.ID               : chr  "B00001" "B00001" "B00001" "B00002" ...
\item $ Sequence.Number        : chr  "0001" "0001" "0001" "0001" ...
\item $ Line.Number            : num  NA NA 1 NA NA 1 NA NA 1 2 ...
\item $ Start.Position         : num  7 NA NA 8 NA NA 7 NA NA NA ...
\item $ Total.Cells.in.Table   : chr  "1 CELL" "" "" "1 CELL" ...
\item $ Total.Cells.in.Sequence: num  NA NA NA 2 NA NA NA NA NA NA ...
\item $ Table.Title            : chr  "UNWEIGHTED SAMPLE COUNT OF THE POPULATION" "Universe:  Total population" "Total" "UNWEIGHTED SAMPLE HOUSING UNITS" ...
\item $ Subject.Area           : chr  "Unweighted Count" "" "" "Unweighted Count" ...
}
For ACS 2008-2012:

\if{html}{\out{<div class="sourceCode">}}\preformatted{length(my.lookup[,1])
#   24741
names(my.lookup)
#  "File.ID"                 "Table.ID"                "Sequence.Number"         "Line.Number"             "Start.Position" \\cr
#  "Total.Cells.in.Table"    "Total.Cells.in.Sequence" "Table.Title"             "Subject.Area"
}\if{html}{\out{</div>}}
}
\description{
Download and read lookup table of information on American Community Survey (ACS) tables, from the Census Bureau,
namely which sequence files on the FTP site contain which tables and which variables.
NOTE: This is largely obsolete now that data(lookup.acs2013) and similar files for other years are in this package.
}
\details{
For information on new Summary File format visit:

https://www.census.gov/programs-surveys/acs/data/summary-file.html

\if{html}{\out{<div class="sourceCode">}}\preformatted{Lookup table of sequence files and variables was here:  \\cr
}\if{html}{\out{</div>}}

\cr
\url{https://www2.census.gov/programs-surveys/acs/summary_file/2017/documentation/user_tools/ACS_5yr_Seq_Table_Number_Lookup.txt}  \cr
\cr\cr
The 2014-2018 folders are here:  \cr
\url{https://www2.census.gov/programs-surveys/acs/summary_file/2018/documentation/geography/}  \cr
\url{https://www2.census.gov/programs-surveys/acs/summary_file/2018/documentation/user_tools/}  \cr
\cr
Data tables by state by seqfile were here:  \cr
\cr
\url{https://www2.census.gov/programs-surveys/acs/summary_file/2017/data/5_year_seq_by_state/Alabama/Tracts_Block_Groups_Only/}  \cr
such as  \cr
\url{https://www2.census.gov/programs-surveys/acs/summary_file/2017/data/5_year_seq_by_state/Alabama/Tracts_Block_Groups_Only/20175al0001000.zip}  \cr
\cr
Geographies in Excel spreadsheets were here:  \cr
\cr
\url{https://www2.census.gov/programs-surveys/acs/summary_file/2017/documentation/geography/5yr_year_geo/}  \cr
such as  \cr
\url{https://www2.census.gov/programs-surveys/acs/summary_file/2017/documentation/geography/5yr_year_geo/ak.xlsx}  \cr
}
\examples{
 \dontrun{
 lookup.acs <- download.lookup.acs(2022)
 }

}
\seealso{
\code{\link[acs:acs.lookup]{acs::acs.lookup()}} which does something similar but is more flexible & robust.
Also see \code{\link[=get.lookup.acs]{get.lookup.acs()}} which does the same without downloading file -- uses the copy in data()
Also see \code{data(lookup.acs2013)} and similar data for other years.
Also see \code{\link[=get_acs_old]{get_acs_old()}}, \code{\link[=get.lookup.file.name]{get.lookup.file.name()}}, \code{\link[=get.url.prefix.lookup.table]{get.url.prefix.lookup.table()}}
}
