% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get.acs.R
\name{get.acs}
\alias{get.acs}
\title{Download Tables from American Community Survey (ACS) 5-year Summary File}
\usage{
get.acs(tables = "B01001", mystates = "all", end.year = "2012",
  base.path = getwd(), data.path = file.path(base.path, "acsdata"),
  output.path = file.path(base.path, "acsoutput"), sumlevel = "both",
  vars = "all", varsfile, new.geo = TRUE, write.files = FALSE,
  save.files = FALSE, write.acspkg = FALSE, testing = FALSE,
  noEditOnMac = FALSE)
}
\arguments{
\item{tables}{Character vector, optional. Defines tables to obtain, such as 'B01001' (the default).
If the user specifies a table called 'ejscreen' then a set of tables used by that tool are included.
Those tables are "B01001", "B03002", "B15002", "B16002", "C17002", "B25034"}

\item{mystates}{Character vector, optional, 'all' by default which means all available states
plus DC and PR but not VI etc. Defines which States to include in downloads of data tables.}

\item{end.year}{Character, optional, '2012' by default. Defines a valid ending year of a 5-year summary file.
Can be '2010' for example. Not all years are tested.}

\item{base.path}{Character, optional, getwd() by default. Defines a base folder to start from, in case
data.path and/or output.path are not specified, in which case a subfolder under base.path,
called acsdata or acsoutput respectively, will be created and used for downloads or outputs of the function.}

\item{data.path}{Character, optional, file.path(base.path, 'acsdata') by default.
Defines folder (created if does not exist) where downloaded files will be stored.}

\item{output.path}{Character, optional, file.path(base.path, 'acsoutput') by default.
Defines folder (created if does not exist) where output files (results of this function) will be stored.}

\item{sumlevel}{Default is "both" (case insensitive) in which case tracts and block groups are returned.
Also c('tract', 'bg') or c(140,150) and similar patterns work.
If "tract" or 140 or some other match, but not block groups, is specified (insensitive to case, tract can be plural or part of a word, etc.),
just tracts are returned.
If "bg" or 150 or "blockgroups" or "block groups" or some other match (insensitive to case, singular or plural or part of a word)
but no match on tracts is specified, just block groups are returned.
Non-matching elements are ignored (e.g., sumlevel=c('bg', 'tracs', 'block') will return block groups
but neither tracts (because of the typo) nor blocks (not available in ACS), with no warning --
No warning is given if sumlevel is set to a list of elements where some are not recognized as matches to bg or tract,
as long as one or more match bg, tracts, or both (or variants as already noted).}

\item{vars}{Optional logical, default is FALSE (in which case all variables from each table will be returned unless otherwise specified -- see below).
This parameter specifies whether to pause and ask the user about which variables are needed in an interactive session in R.
This gives the user a chance to prepare the file "variables needed.csv" (or just ensure it is ready),
or to edit and save "variables needed.csv" within a window in the default editor used by R (the user is asked which of these is preferred).
If vars=FALSE, the function just looks in \code{data.folder} for a file called "variables needed.csv" that, if used,
must specify which variables to keep from each table.
The format of that file should be the same as is found in the file "variables needed template.csv" created by this function --
keeping the letter "Y" in the column named "keep" indicates the given variable is needed.
A blank or "N" or other non-Y specifies the variable should be dropped and not returned by get.acs().
If the "variables needed.csv" file is not found, however, this function looks for and uses the file called "variables needed template.csv"
which is written by this function and specifies all of the variables from each table, so all variables will be retained and returned by get.acs().}

\item{varsfile}{See help for \code{\link{set.needed}} for details. Optional name of file that can be used to specify which variables are needed from specified tables.
If varsfile is specified, parameter vars is ignored, and the function just looks in folder for file called filename, e.g., "variables needed.csv"
that should specify which variables to keep from each table.}

\item{new.geo}{Default is TRUE. If FALSE, just uses existing downloaded geo file if possible. If TRUE, forced to download geo file even if already done previously.}

\item{write.files}{Default is FALSE, but if TRUE then data-related csv files are saved locally -- Saves longnames, full fieldnames as csv file, in working directory.}

\item{save.files}{Default is FALSE, but if TRUE then various intermediate image files are saved as .RData files locally in working directory.}

\item{write.acspkg}{Default is FALSE. If TRUE, saves csv file of tracts and file of block groups, for each of the \code{tables},
in a format that the Census Bureau American Fact Finder provides as downloadable tables. That format can be easily read in by the very useful \pkg{acs} package.}

\item{testing}{Default is FALSE, but if TRUE more information is shown on progress, using cat() and while downloading, and more files (csv) are saved in working directory.}

\item{noEditOnMac}{FALSE by default. If TRUE, do not pause to allow edit() to define which variables needed from each table,
  when on Mac OSX, even if vars=TRUE. Allows you to avoid problem in RStudio if X11 not installed.}
}
\value{
By default, returns a list of ACS data tables and information about them, with these elements in the list: \cr
  bg, tracts, headers, and info. The headers and info elements are data.frames providing metadata such as short and long field names.
  The same column names are found in x$info and x$headers, but headers has more rows. The info table just provides information about each
  data variable in the estimates tables. The headers table provides similar information but made to match the bg or tract format,
  so the headers table has as many rows as bg or tracts has columns -- enough for the estimates and MOE fields, and the basic fields such as FIPS.
  The info data.frame can look like this, for example: \cr

 \code{
  'data.frame':	xxxx obs. of  9 variables:  \cr
 $ table.ID       : chr  "B01001" "B01001" "B01001" "B01001" ...  \cr
 $ line           : num  1 2 3 4 5 6 7 8 9 10 ...  \cr
 $ shortname      : chr  "B01001.001" "B01001.002" "B01001.003" "B01001.004" ...  \cr
 $ longname       : chr  "Total:" "Male:" "Under 5 years" "5 to 9 years" ...  \cr
 $ table.title    : chr  "SEX BY AGE" "SEX BY AGE" "SEX BY AGE" "SEX BY AGE" ...  \cr
 $ universe       : chr  "Universe:  Total population" "Universe:  Total population" "Universe:  Total population" "Universe:  Total population" ...  \cr
 $ subject        : chr  "Age-Sex" "Age-Sex" "Age-Sex" "Age-Sex" ...  \cr
 $ longname2      : chr  "Total" "Male" "Under5years" "5to9years" ...  \cr
 $ longname.unique: chr  "Total:|SEX BY AGE" "Male:|SEX BY AGE" "Under 5 years|SEX BY AGE" "5 to 9 years|SEX BY AGE" ...  \cr
 }
}
\description{
This function will download and parse 1 or more tables of data from the American Community Survey's
  5-year Summary File FTP site, for all Census tracts and/or block groups in specified State(s).
  Estimates and margins of error are obtained, as well as long and short names for the variables,
  which can be specified if only parts of a table are needed.
}
\details{
The United States Census Bureau provides detailed demographic data by US Census tract and block group
  in the American Community Survey (ACS) 5-year summary file via their FTP site. For those interested in block group or tract data,
  Census Bureau tools tend to focus on obtaining data one state at a time rather than for the entire US at once.
  The \code{get.acs} function lets a user specify (tables and) variables needed. The code will look up what sequence files
  contain those tables. Using a table of variables for those selected tables, a user can specify variables
  or tables to drop by putting x or anything other than "Y" in the column specifying needed variables. \cr\cr
  For ACS documentation, see for example:\cr\cr
\url{http://www.census.gov/programs-surveys/acs.html} \cr
or\cr
\url{http://www.census.gov/acs/www/Downloads/data_documentation/SumFileTemp/2012-5-Year/Sequence_Number_and_Table_Number_Lookup.xls}\cr
linked from\cr
\url{http://www.census.gov/programs-surveys/acs/data/summary-file.html} \cr \cr
#####################################	ADDITIONAL NOTES ######################################\cr
	NOTES ON NON-NUMERIC FIELDS IN ESTIMATE AND MOE FILES:\cr \cr
FROM 2008-2012 ACS 5-Year Summary File Technical Documentation:\cr \cr
Some data values represent unique situations where either the information to be conveyed is
an explanation for the absence of data, represented by a #symbol in the data display, such as "(X)",
or the information to be conveyed is an open-ended distribution, such as 115 or greater, represented by 115+.
The following special data values can appear in the ACS Summary File table as an explanation for the absence of data:\cr
2008-2012 ACS 5-Year Summary File Technical Documentation\cr \cr
? Missing Value = ??    **********\cr
A missing string indicates that the estimate is unavailable. (This appears in the estimates and margins or error
files as two commas adjacent to each #other without anything between them, or if the last cell in a data file is
filtered then you get a comma followed immediately by a carriage return or #EOF.)
A missing value indicates when an estimate is missing because of filtering for geographic restrictions,
coefficients of variations (CV), or was #removed due to the Disclosure Review Board?s (DRB) requirements.
For more detail on filtering, please see Appendix C.5.\cr \cr

? Dot = ?.?		********\cr
A dot indicates when the estimate has no sample observations or too few sample observations.
In the margin of error files, this value could also indicate #that the margin of error is unavailable for
a median estimate that has been replaced with a jam value.\cr \cr

? Zero = ?0?\cr
A ?0? entry in the margin of error column indicates that the estimate is controlled.
A statistical test for sampling variability is not appropriate.
This #is similar to the ?*****? symbol used in American FactFinder.\cr \cr

? Negative 1 = ?-1?\cr
This indicates that an estimate does not contain a Margin of Error.
Tables B00001, B00002, and tables starting with B98 and B99 do not have margin of #error (MOE)
associated with them. The MOE calculations are set to -1 for these tables.\cr\cr

? Jam Values for Medians\cr
The following is a listing of the jam values for medians. For example, if there is an estimate of "2499"
for table B10010, then it does not indicate a #dollar amount. It means that the median is somewhere
below 2,500 and thus isn't calculated. \cr\cr

"Jam Value","Actual Meaning","Use for Medians" \cr
"0","1 or less","Age, Duration of Marriage" \cr
"9","9.0 or more","Rooms" \cr
"10","10.0 or less","Gross Rent as Percentage of Income, Owner Costs as Percentage of Income" \cr
"50","50.0 or more","Gross Rent as Percentage of Income, Owner Costs as Percentage of Income" \cr
"99","100 or less","Rent, Gross Rent, Selected Monthly Owner Costs, Monthly Housing Costs" \cr
"101","101 or more","Duration of Marriage" \cr
"116","115 or more","Age" \cr
"199","200 or less","Tax" \cr
"1001","1,000 or more","Selected Monthly Owner Costs" \cr
"1939","1939 or earlier","Year Built" \cr
"1969","1969 or earlier","Year Moved In" \cr
"2001","2,000 or more","Rent, Gross Rent" \cr
"2005","2005 or later","Year Built, Year Moved In" \cr
"2499","2,500 or less","Income, Earnings" \cr
"4001","4,000 or more","Selected Monthly Owner Costs, Monthly Housing Costs" \cr
"9999","10,000 or less","Value" \cr
"10001","10,000 or more","Tax" \cr
"200001","200,000 or more","Income" \cr
"250001","250,000 or more","Income, Earnings" \cr
"1000001","1,000,000 or more","Value" \cr
 ############################################### \cr \cr

   in older 2005-2009 and 2005-2010 also these were the sequence numbers, but they change over time: \cr
seqnum <- "0010" has ageunder5m = B01001.003 etc. \cr
seqnum <- "0013" has hisp = B03002.012 etc. \cr
seqnum <- "0040" has age25up = B15002.001 etc. \cr
seqnum <- "0042" has lingisospanish = B16002.004 etc. \cr
seqnum <- "0046" has povknownratio = C17002.001 etc. \cr
NOTE:  seq file 98 was used for year built in ACS 2005-2009 but in ACS 2006-2010 that is in seq file 97 \cr
############# \cr \cr

#####################################################################################  \cr
	NOTES on where to obtain ACS data - sources for downloads of summary file data  \cr
##################################################################################### \cr\cr \cr

FOR ACS SUMMARY FILE DOCUMENTATION, SEE \cr
\url{http://www.census.gov/acs/www/data_documentation/summary_file/} \cr \cr

	As of 12/2012, ACS block group/tract summary file ESTIMATES on FTP site is provided as either \cr\cr

 LARGER THAN NECESSARY: \cr\cr

	*** all states and all tables in one huge tar.gz file (plus a zip of all geography codes), \cr
	\url{ftp://ftp.census.gov/acs2011_5yr/summaryfile//2007-2011_ACSSF_All_In_2_Giant_Files(Experienced-Users-Only)} \cr
	\url{ftp://ftp.census.gov/acs2011_5yr/summaryfile//2007-2011_ACSSF_All_In_2_Giant_Files(Experienced-Users-Only)/2011_ACS_Geography_Files.zip} \cr
	2011_ACS_Geography_Files.zip		# \cr
	Tracts_Block_Groups_Only.tar.gz		# (THIS HAS ALL THE SEQUENCE FILES FOR EACH STATE IN ONE HUGE ZIP FOLDER) \cr\cr

or \cr\cr

 MORE FOCUSED DOWNLOADS: \cr\cr

	*** one zip file per sequence file PER STATE:    (plus csv of geography codes) \cr \cr

	\url{http://www2.census.gov/acs2011_5yr/summaryfile/2007-2011_ACSSF_By_State_By_Sequence_Table_Subset/DistrictOfColumbia/Tracts_Block_Groups_Only/} \cr
        20115dc0002000.zip (Which has the e file and m file for this sequence file in this state) \cr
        g20115dc.csv  (which lets you link FIPS to data) \cr
	so this would mean 50+ * about 7 sequence files? = about 350 zip files?? each with 2 text files. Plus 50+ geo csv files.
	so downloading about 400 files and expanding to about 750 files, and joining into one big file.\cr

OR \cr\cr

 PREJOINED TO TIGER BLOCK GROUP BOUNDARIES SHAPEFILES/ GEODATABASES - \cr
	ONE PER STATE HAS SEVERAL TABLES BUT NOT B16001, B16002 (many languages but tracts only), B16004 (has block groups but fewer languages) \cr
	\url{http://www.census.gov/geo/maps-data/data/tiger-data.html} \cr
 BUT NOT one file per sequence file FOR ALL STATES AT ONCE. \cr
Estimates & margin of error (MOE), (ONCE UNZIPPED), and GEOgraphies (not zipped) are in 3 separate files. \cr \cr

also, data for entire US for one seq file at a time, but not tract/bg -- just county and larger? -- is here, e.g.:
\url{ftp://ftp.census.gov/acs2011_1yr/summaryfile//2011_ACSSF_By_State_By_Sequence_Table_Subset/UnitedStates/20111us0001000.zip} \cr
GEO files:\cr
Note the US file is not bg/tract level:  geo for whole US at once doesn't have tracts and BGs
\url{ftp://ftp.census.gov/acs2011_1yr/summaryfile//2011_ACSSF_By_State_By_Sequence_Table_Subset/UnitedStates/g20111us.csv} \cr\cr

OTHER SOURCES include \cr
\itemize{
  \item \pkg{acs} package for R - very useful for modest numbers of Census units rather than every block group in US
  \item \url{http://www.NHGIS.org} - very useful for block group (or tract/county/state/US) datasets
  \item DataFerrett (\url{http://dataferrett.census.gov/AboutDatasets/ACS.html}) -- not all tracts in US at once
  \item American Fact Finder (\url{http://www.census.gov/acs/www/data/data-tables-and-tools/american-factfinder/}) (not block groups for ACS SF, and the tracts are not for the whole US at once)
  \item ESRI - commercial
  \item Geolytics - commercial
  \item etc.
}
}
\examples{
\dontrun{
  t( get.table.info(c('B01001', 'C17002', 'B03002')) ) # Basic info on ACS tables
  out <- get.acs(mystates=c('dc','de')) # Data for just DC & DE, just the default table.
  names(out$bg); cat('\\n\\n'); head(out$info)
  cbind(longname=out$info$longname,
        total=colSums(out$bg[ , names(out$bg) \%in\% out$info$shortname ]))
  # to see data on 2 places, 1 per column, with short and long field names
  cbind( out$headers$longname, t(out$bg[1:2, ]) )
  # to see 7 places, 1 per row, with short and long field name as header
  head( rbind(out$headers$longname, out$bg) )[,1:7]
  out <- get.acs(mystates='de', tables=c('B01001', 'C17002'))  # just 2 tables for just Delaware
  head(out$info); head(out$bg)
  # uses all EJSCREEN defaults and the specified folders:
  out <- get.acs(base.path='~', data.path='~/ACS/temp', output.path='~/ACS/results')
  head(out$info); head(out$bg)
  out <- get.acs(tables=c('ejscreen', 'B16001')) # all tables needed for EJSCREEN,
    plus 'B16001', with variables specified in 'variables needed.csv', all states&DC &PR?
  head(out$info); head(out$bg)
 }
}
\seealso{
\pkg{acs} package which allows you to download and work with ACS data (using the API and your own key).
}

