% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get.datafile.prefix.R
\name{get.datafile.prefix}
\alias{get.datafile.prefix}
\title{Get first part of ACS datafile name.}
\usage{
get.datafile.prefix(end.year = "2012")
}
\arguments{
\item{end.year}{Optional character, default is "2012", specifying last year of 5-year summary file data.}
}
\value{
By default, returns "20125"
}
\description{
Get the first part of the datafile name for the ACS 5-year summary file datafiles on the US Census Bureau FTP site.
}
\seealso{
\code{\link{get.acs}}, \code{\link{datafile}}, \code{\link{geofile}}, \code{\link{get.zipfile.prefix}}
}

