% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get.field.info.R
\name{get.field.info}
\alias{get.field.info}
\title{Get short and long field names etc for ACS tables}
\usage{
get.field.info(tables, end.year = "2012", table.info.only = FALSE,
  moe = FALSE)
}
\arguments{
\item{tables}{Required vector of tables such as "B01001"}

\item{end.year}{Last year of 5-year summary file such as '2012' (default)}

\item{table.info.only}{FALSE by default.}

\item{moe}{FALSE by default. Margin of error variables also included if TRUE,
but their names are identical to those of estimates fields other than being MOE instead of estimate.}
}
\value{
data.frame of information about each table and each variable in table: \cr
  Value returned is data.frame of info about each table and also each variable in the table\cr
  e.g., longname2 which is version where no spaces or colons or escaped quotation marks etc. \cr
}
\description{
Get info on tables from US Census Bureau for American Community Survey 5-year summary file.
}
\examples{
\dontrun{
  finfo <- get.field.info(c('B17020A', 'B17020H'))
  cbind( names(tracts), substr(finfo$longname.unique[match(names(tracts), finfo$shortname)], 1, 100))
  foundnames <- names(tracts)[names(tracts) \%in\% finfo$shortname]
  cbind( foundnames, substr(finfo$longname.unique[match(foundnames, finfo$shortname)], 1, 100))
}
}
\seealso{
\code{\link{get.acs}}, \code{\link{get.table.info}}, \code{\link{get.table.info2}} and \code{\link{get.field.info}}
}

