% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get.lookup.acs.R
\name{get.lookup.acs}
\alias{get.lookup.acs}
\title{Download File with Information about ACS 5-Year Summary File Tables}
\usage{
get.lookup.acs(end.year = "2012", folder = getwd())
}
\arguments{
\item{end.year}{Character, optional, '2012' by default, which specifies the 2008-2012 dataset.
Defines which 5-year summary file to use, based on end-year.}

\item{folder}{Optional path to where to download file to, defaults to current working directory.}
}
\value{
By default, returns lookup.acs if that was already in memory, but typically returns
  a data.frame with these fields:
  \itemize{
    \item Table ID
    \item Sequence Number
    \item Line Number
    \item Start Position
    \item Total Cells in Table
    \item Total Cells in Sequence
    \item Table Title
    \item Subject Area
  }
  \cr
  For ACS 2008-2012: \cr
  length(my.lookup[,1]) \cr
  [1] 24741\cr
  names(my.lookup)\cr
  [1] "File.ID"                 "Table.ID"                "Sequence.Number"         "Line.Number"             "Start.Position"\cr
  [6] "Total.Cells.in.Table"    "Total.Cells.in.Sequence" "Table.Title"             "Subject.Area"\cr
}
\description{
Download and read lookup table of information on American Community Survey (ACS) tables, from the Census Bureau,
  namely which sequence files on the FTP site contain which tables and which variables.
}
\details{
Note that lookup may already be in memory from prior work and is not downloaded again if named lookup.acs.
  Code was developed late 2013 through mid 2015.
  \cr\cr
  The source of this lookup table is, for example,
  \url{ftp://ftp.census.gov/acs2012_5yr/summaryfile/Sequence_Number_and_Table_Number_Lookup.txt}
  \cr\cr
  Note: The \pkg{acs} package provides a more powerful version of this, called \code{\link[acs]{acs.lookup}}
}
\examples{
\dontrun{
 lookup.acs <- get.lookup.acs()
 }
}
\seealso{
\code{\link[acs]{acs.lookup}} which does something similar but is more flexible & robust.
  Also see \code{data(lookup.acs)}. Also see \code{\link{get.acs}} which uses this.
}

