% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.read.geo.R
\name{get.read.geo}
\alias{get.read.geo}
\title{Download (if necessary) and merge GEO files for ACS}
\usage{
get.read.geo(
  mystates,
  new.geo = FALSE,
  folder = getwd(),
  end.year = acsdefaultendyearhere,
  testing = FALSE,
  silent = FALSE
)
}
\arguments{
\item{mystates}{Character vector of 2-character state abbreviations, required.}

\item{new.geo}{Logical value, optional, FALSE by default. If FALSE, if geo exists in memory don't download and parse again.}

\item{folder}{Defaults to current working directory.}

\item{end.year}{optional character year to specify last year of 5-year summary file.}

\item{testing}{Default to FALSE. If TRUE, provides info on progress of download.}

\item{silent}{Default is FALSE.}
}
\value{
Returns a data.frame of all states geo info.

ACS 2008-2012, tract and block group counts:

table(geo$SUMLEVEL)

\if{html}{\out{<div class="sourceCode">}}\preformatted{140    150
}\if{html}{\out{</div>}}

74001 220333
}
\description{
Returns a data.frame of all states merged geo info and saves geo.RData to working directory.
}
\details{
Uses download.geo() then read.geo(), and then does some cleanup.

Note that if this finds the geographic file in folder already, it will not download it again even if that file was corrupt.
Read and compile geo data for entire USA with PR DC,

This takes some time for the entire USA:

2 to 10 minutes, depending.

Remaining fields in geo:

"STUSAB"   "SUMLEVEL" "GEOID"    "FIPS"     "KEY"

NOTE: do not really need GEOID or KEY.

GEOID is redundant, but might be useful for joining to shapefiles/ boundaries

Also, could specify here if "NAME" field from geo files should be dropped -
it might be useful but takes lots of RAM and encoding of S
panish characters in Puerto Rico caused a problem in Mac OSX.

NOTE FROM CENSUS:

The ACS Summary File GEOID contains the necessary information to
connect to the TIGER/Line Shapefiles,
but it needs to be modified in order to exactly match up.
Notice that the ACS GEOID, 05000US10001, contains
the TIGER/Line GEOID string, 10001.
In order to create an exact match of both GEOIDs,
it is necessary to
remove all of the characters before and including
the letter S in the ACS Summary File.
By removing these characters, the new GEOID in the
ACS Summary File exactly matches the field GEOID in
the TIGER/Line Shapefiles.
}
\seealso{
\code{\link[=get_acs_old]{get_acs_old()}} which uses this, and \code{\link[=download.geo]{download.geo()}}
}
