% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.read.geo.R
\name{get.read.geo}
\alias{get.read.geo}
\title{Download (if necessary) and merge GEO files for ACS}
\usage{
get.read.geo(
  mystates,
  new.geo = FALSE,
  folder = getwd(),
  end.year = acsdefaultendyearhere_func(),
  testing = FALSE,
  silent = FALSE
)
}
\arguments{
\item{mystates}{Character vector of 2-character state abbreviations, required.}

\item{new.geo}{Logical value, optional, FALSE by default. If FALSE, if geo exists in memory don't download and parse again.}

\item{folder}{Defaults to current working directory.}

\item{end.year}{optional character year to specify last year of 5-year summary file.}

\item{testing}{Default to FALSE. If TRUE, provides info on progress of download.}

\item{silent}{Default is FALSE.}
}
\value{
Returns a data.frame of all states geo info. \cr
  # FOR ACS 2008-2012, tract and block group counts: \cr
  table(geo$SUMLEVEL) \cr
    140    150  \cr
  74001 220333  \cr

 Remaining fields in geo: \cr
 "STUSAB"   "SUMLEVEL" "GEOID"    "FIPS"     "KEY" \cr
 NOTE: do not really need GEOID or KEY. \cr
 GEOID is redundant, but might be useful for joining to shapefiles/ boundaries \cr
 Also, could specify here if "NAME" field from geo files should be dropped -
 it might be useful but takes lots of RAM and encoding of Spanish characters in Puerto Rico caused a problem in Mac OSX. \cr
 NOTE FROM CENSUS:  \cr
 The ACS Summary File GEOID contains the necessary information to connect to the TIGER/Line Shapefiles,
 but it needs to be modified in order to exactly match up.
 Notice that the ACS GEOID, 05000US10001, contains the TIGER/Line GEOID string, 10001.
 In order to create an exact match of both GEOIDs, it is necessary to
 remove all of the characters before and including the letter S in the ACS Summary File.
 By removing these characters, the new GEOID in the ACS Summary File exactly matches the field GEOID in the TIGER/Line Shapefiles.
}
\description{
Returns a data.frame of all states merged geo info and saves geo.RData to working directory.
}
\details{
Uses download.geo() then read.geo(), and then does some cleanup.

Note that if this finds the geographic file in folder already, it will not download it again even if that file was corrupt.
Read and compile geo data for entire USA with PR DC, \cr
This takes some time for the entire USA:  \cr
2 minutes 48 seconds on \cr
MacBook Pro 16 GB RAM, SSD, 2.7 GHz Intel Core i7\cr
R version: R 3.0.2 GUI 1.62 Snow Leopard build (6558)\cr
8 minutes 20 seconds on \cr
R version 3.0.1 \cr
Dell Latitude E6400, x86 Family 6 Model 23 Stepping 6 GenuineIntel ~2260 Mhz, \cr
Total Physical Memory 4,096.00 MB, Available Physical Memory 1.97 GB, Total Virtual Memory 2.00 GB, Available Virtual Memory 1.94 GB, \cr
Microsoft Windows XP Professional \cr
############## \cr
}
\seealso{
\code{\link{get.acs}} which uses this, and \code{\link{download.geo}}
}
