% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.table.info.R
\name{get.table.info}
\alias{get.table.info}
\title{Get field names etc for ACS tables}
\usage{
get.table.info(tables, end.year = "2019", table.info.only = TRUE, moe = FALSE)
}
\arguments{
\item{tables}{Required vector of tables such as "B01001"}

\item{end.year}{Last year of 5-year summary file such as '2012'}

\item{table.info.only}{TRUE by default. See \code{\link{get.field.info}}}

\item{moe}{FALSE by default. If TRUE, returns MOE versions of field names and descriptions.}
}
\value{
data.frame of information about each table and each variable in table: \cr
  Table.ID, Line.Number, Table.Title, table.var, varname2 \cr\cr
    # Value returned is data.frame of info about each table and also each variable in the table:\cr

        Table.ID Line.Number                                                 Table.Title  table.var                                           varname2\cr
  7       B01001          NA                                                  SEX BY AGE       <NA>                                           SEXBYAGE\cr
  8       B01001          NA                                 Universe:  Total population       <NA>                            UniverseTotalpopulation\cr
  9       B01001           1                                                      Total: B01001.001                                              Total\cr
  10      B01001           2                                                       Male: B01001.002                                               Male\cr
}
\description{
Get info on tables from US Census Bureau for American Community Survey 5-year summary file.
}
\details{
Wrapper for \code{\link{get.table.info2}} which is a wrapper for \code{\link{get.field.info}}
}
\seealso{
\code{\link{get.acs}}, \code{\link{get.table.info}}, and \code{\link{get.field.info}}
}
