% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get.table.info2.R
\name{get.table.info2}
\alias{get.table.info2}
\title{Get field names etc for ACS tables}
\usage{
get.table.info2(tables, end.year = "2012", table.info.only = TRUE)
}
\arguments{
\item{tables}{Required vector of tables such as "B01001"}

\item{end.year}{Last year of 5-year summary file such as '2012' (default)}

\item{table.info.only}{TRUE by default. See \code{\link{get.field.info}}}
}
\value{
data.frame of information about each table and each variable in table.
}
\description{
Get info on tables from US Census Bureau for American Community Survey 5-year summary file.
}
\seealso{
\code{\link{get.acs}}, \code{\link{get.table.info}}, \code{\link{get.table.info2}} and \code{\link{get.field.info}}
}

