% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.tracts.R
\name{get.tracts}
\alias{get.tracts}
\title{Get just the Census tracts part of existing ACS data}
\usage{
get.tracts(merged.tables.mine)
}
\arguments{
\item{merged.tables.mine}{Required set of tables in format used by \code{\link[=get_acs_old]{get_acs_old()}}}
}
\value{
subset of the inputs, same format
}
\description{
Helper function to return just the tract resolution part of a dataset in \code{\link[=get_acs_old]{get_acs_old()}}
}
\seealso{
\code{\link[=get_acs_old]{get_acs_old()}}, \code{\link[=get.bg]{get.bg()}}
}
