% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.url.prefix.lookup.table.R
\name{get.url.prefix.lookup.table}
\alias{get.url.prefix.lookup.table}
\title{Get first part of name(s) of URL(s) for ACS 5-year summary file data}
\usage{
get.url.prefix.lookup.table(end.year = "2012")
}
\arguments{
\item{end.year}{Optional end year for 5-year summary file, as character, defaulting to "2012"}
}
\value{
Returns character element that is first part of URL such as "ftp://ftp.census.gov/acs2012_5yr/summaryfile/"
}
\description{
Returns first part of URL(s) of folders (on Census Bureau FTP site) with zip file(s) based on end year.
}
\seealso{
\code{\link{get.acs}}, \code{\link{get.lookup.acs}}, \code{\link{get.lookup.file.name}}
}

