% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.zipfile.prefix.R
\name{get.zipfile.prefix}
\alias{get.zipfile.prefix}
\title{Get first part of ACS zip file name.}
\usage{
get.zipfile.prefix(end.year = acsdefaultendyearhere_func())
}
\arguments{
\item{end.year}{Optional character, specifying last year of 5-year summary file data.}
}
\description{
Get the first part of the zipfile name for the ACS 5-year summary file datafiles on the US Census Bureau FTP site.
}
\seealso{
\code{\link{get.acs}}, \code{\link{get.datafile.prefix}}, \code{\link{datafile}}, \code{\link{geofile}}, \code{\link{get.zipfile.prefix}}
}
