% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getseqnumsviafilenames.R
\name{getseqnumsviafilenames}
\alias{getseqnumsviafilenames}
\title{Infer Sequence file numbers based on ACS 5-yr filenames in folder}
\usage{
getseqnumsviafilenames(folder = getwd())
}
\arguments{
\item{folder}{Default is current working directory.}
}
\value{
Returns a character vector of unique sequence file numbers
}
\description{
Helper function to look for unzipped csv files of estimates for American Community Survey (ACS)
  5-year summary file data obtained from US Census FTP site, based on pattern matching, and infer seqfile numbers
  based on those filenames.
}
\seealso{
\code{\link{read.concat.states}} which uses this
}

