% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join.geo.to.tablist.R
\name{join.geo.to.tablist}
\alias{join.geo.to.tablist}
\title{Join US ACS data and US geo files on FIPS}
\usage{
join.geo.to.tablist(
  mygeo,
  my.list.of.tables,
  save.csv = FALSE,
  sumlevel = "both",
  folder = getwd(),
  testing = FALSE,
  end.year = acsdefaultendyearhere_func()
)
}
\arguments{
\item{mygeo}{Required geo file. See \code{\link[=get.acs]{get.acs()}} and \code{\link[=get.read.geo]{get.read.geo()}}}

\item{my.list.of.tables}{List of data tables resulting from prior steps in \code{\link[=get.acs]{get.acs()}}}

\item{save.csv}{FALSE by default. Specifies whether to save each data table as csv format file.}

\item{sumlevel}{Default is "both", specifies if "tracts" or "blockgroups" or "both" should be used.}

\item{folder}{Default is current working directory.}

\item{testing}{Default is FALSE. If TRUE, prints more information.}

\item{end.year}{Default is "" -- used in naming file if save.csv=TRUE}
}
\value{
Returns a list of data.frames, where each element of the list is one ACS table, such as table B01001.
}
\description{
Read the processed csv files of estimates and MOE (margin of error) for American Community Survey (ACS)
5-year summary file data obtained from US Census FTP site, and join with geographic information from geo file.
}
\seealso{
\code{\link[=get.acs]{get.acs()}} and \code{\link[=get.read.geo]{get.read.geo()}}
}
