% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookup.acs2021.R
\docType{data}
\name{lookup.acs2021}
\alias{lookup.acs2021}
\title{ACS_5yr_Seq_Table_Number_Lookup.txt for ACS dataset}
\description{
ACS_5yr_Seq_Table_Number_Lookup.txt for ACS dataset
}
\details{
Created using script in data-raw folder of source package.
}
\seealso{
\link{lookup.acs}
}
