% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/nhgis.R
\name{nhgis}
\alias{nhgis}
\title{Read and Parse NHGIS.org ACS Data Files and Codebooks}
\usage{
nhgis(base.path = getwd(), code.dir = file.path(base.path, "nhgiscode"),
  data.dir = file.path(base.path, "nhgisdata"), silent = FALSE,
  savefiles = FALSE)
}
\arguments{
\item{base.path}{Optional base path, default is getwd()}

\item{code.dir}{Optional path where extra code is. Not used.}

\item{data.dir}{Optional path where data files are stored and output could be saved. Default is nhgiscode folder under base.path}

\item{silent}{Optional, FALSE by default, whether to print info about progress, filenames found, etc.}

\item{savefiles}{Optional, FALSE by default, whether to save .RData and maybe csv files of output returned.}
}
\value{
Returns a list of data tables.
}
\description{
Read downloaded and unzipped csv and txt files obtained from NHGIS.org,
  with US Census Bureau data from the American Community Survey (ACS).
}
\details{
This is designed to read and parse csv and txt files
  obtained from NHGIS.org and already unzipped in a local folder.
  Obtaining NHGIS.org data requires an account at
  \url{https://data2.nhgis.org/main}, \url{https://www.nhgis.org}
  Data can be downloaded by selecting, for example, \cr
  tracts and block groups, all in US, acs2007-2011, and specifying the desired ACS Table(s). \cr
  Research using NHGIS data should cite it as: \cr
  Minnesota Population Center. National Historical Geographic Information System: Version 2.0. Minneapolis, MN: University of Minnesota 2011.
}
\seealso{
\code{\link{nhgisread}}, \code{\link{get.acs}}, \code{\link{get.datafile.prefix}}, \code{\link{datafile}}, \code{\link{geofile}}, \code{\link{get.zipfile.prefix}}
}

