% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhgisreadcodebook.R
\name{nhgisreadcodebook}
\alias{nhgisreadcodebook}
\title{Read NHGIS.org ACS Codebook File}
\usage{
nhgisreadcodebook(codebookfile, folder = getwd())
}
\arguments{
\item{codebookfile}{Name(s) of codebook file(s).}

\item{folder}{Optional path where files are found. Default is getwd()}
}
\value{
Returns a named list: data, contextfields, fields, tables, geolevel, years, dataset
}
\description{
Helper function used by \code{\link{nhgis}} to read downloaded and unzipped codebook files obtained from NHGIS.org,
  for US Census Bureau data from the American Community Survey (ACS).
}
\seealso{
\code{\link{nhgis}} which uses this, \code{\link{nhgisread}} for reading datafiles, \code{\link{get.acs}}, \code{\link{get.datafile.prefix}}, \code{\link{datafile}}, \code{\link{geofile}}, \code{\link{get.zipfile.prefix}}
}
