% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/read.concat.states.R
\name{read.concat.states}
\alias{read.concat.states}
\title{Read and concatenate State files of ACS data}
\usage{
read.concat.states(tables, mystates, geo, needed, folder = getwd(),
  end.year = "2012", save.files = TRUE, sumlevel = "both",
  testing = FALSE)
}
\arguments{
\item{tables}{Character vector of table numbers needed such as 'B01001'}

\item{mystates}{Character vector of 2-character state abbreviations. Default is all states.}

\item{geo}{Required table of geographic identifiers to merge with data here. See \code{\link{get.read.geo}}}

\item{needed}{Required table specifying which variables to use from each table. See \code{\link{set.needed}} and \code{\link{get.acs}}}

\item{folder}{Default is current working directory.}

\item{end.year}{Default is "2012", specifies end year of 5-year summary file.}

\item{sumlevel}{Default is "both", specifies if "tracts" or "blockgroups" or "both" should be returned.}

\item{testing}{Default is FALSE. If TRUE, prints filenames but does not unzip them.}
}
\value{
Returns a list of data.frames, where each element of the list is one ACS table, such as table B01001.
}
\description{
Read the unzipped csv files of estimates and MOE (margin of error) for American Community Survey (ACS)
  5-year summary file data obtained from US Census FTP site.
  These State-specific csv files are combined into a single national result for each Census data table.
}
\details{
This is a slow function using rbind() -- It can takes several minutes for a national dataset of a few tables.
  Also note some sequence files might have >1 table in the sequence file,
  so must loop through tables not seqfiles to get start points right.
}
\seealso{
\code{\link{get.acs}}
}

