% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.concat.states.R
\name{read.concat.states}
\alias{read.concat.states}
\title{Read and concatenate State files of ACS data}
\usage{
read.concat.states(tables, mystates, geo, needed, folder = getwd(),
  output.path, end.year = "2012", save.files = TRUE, sumlevel = "both",
  testing = FALSE, dt = TRUE, silent = FALSE)
}
\arguments{
\item{tables}{Optional character vector of table numbers needed such as 'B01001', but default is all tables from each sequence file found.}

\item{mystates}{Optional character vector of 2-character state abbreviations. Default is all states for which matching filenames are found in folder.}

\item{geo}{Optional table of geographic identifiers that elsewhere would be merged with data here.
If provided, can uses it here to look up data file length, based on state abbrev's list. See \code{\link{get.read.geo}}
If geo is not provided, the function still reads each file whatever its length.}

\item{needed}{Optional data.frame specifying which variables to keep from each table. Default is to keep all. See \code{\link{set.needed}} and \code{\link{get.acs}}}

\item{folder}{Default is current working directory. Specifies where the csv files are to be found.}

\item{output.path}{Default is whatever the parameter \code{folder} is set to. Results as .RData files are saved here if save.files=TRUE.}

\item{end.year}{Default is "2012", specifies end year of 5-year summary file.}

\item{save.files}{Default is TRUE, in which case it saves each resulting table/ data file on disk.}

\item{sumlevel}{Default is "both". Specifies if "tracts" or "blockgroups" or "both" should be returned.}

\item{testing}{Default is FALSE. If TRUE, prints filenames but does not unzip them, and prints more messages.}

\item{dt}{Optional logical, TRUE by default, specifies whether data.table::fread should be used instead of read.csv}

\item{silent}{Default is FALSE. Whether to send progress info to standard output.}
}
\value{
Returns a list of data.frames, where each element of the list is one ACS table, such as table B01001.
}
\description{
Read the unzipped csv files of estimates and MOE (margin of error) for American Community Survey (ACS)
  5-year summary file data obtained from US Census FTP site.
  These State-specific csv files are combined into a single national result for each Census data table.
}
\details{
This can use read.csv (takes about 1 minute total for one table, all states, est and moe).
  It can use data.table::fread (default method), which is faster.
}
\seealso{
\code{\link{get.acs}}
}
