% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/set.needed.R
\name{set.needed}
\alias{set.needed}
\title{Specify which ACS Variables are Needed}
\usage{
set.needed(tables, lookup.acs)
}
\arguments{
\item{tables}{Character vector, required. Specifies which ACS tables.}

\item{lookup.acs}{Data.frame, optional. Defines which variables are in which tables. Output of \code{\link{get.lookup.acs}}}
}
\value{
Returns data.frame of info on which variables are needed from each table, much like annotated version of lookup.acs.
}
\description{
Utility used by get.acs to help user specify which variables are needed.
  User can specify this in a file in the working directory, modifying "variables needed template.csv"
  that this function can create based on tables parameter, to create user-defined "variables needed.csv"
}
\seealso{
\code{\link{get.acs}} which uses this
}

