% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/unzip.datafiles.R
\name{unzip.datafiles}
\alias{unzip.datafiles}
\title{Unzip ACS datafile per state, downloading missing ones first}
\usage{
unzip.datafiles(tables, mystates, folder = getwd())
}
\arguments{
\item{tables}{Character vector of table numbers needed such as 'B01001'}

\item{mystates}{Character vector of 2-character state abbreviations. Default is all states.}

\item{folder}{Default is current working directory.}
}
\value{
Side effect is unzipping file on disk
}
\description{
Unzip ACS datafile for each specified US State, extracting specified table(s), downloading missing zip files first.
}
\seealso{
\code{\link{get.acs}}
}

