% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/url.to.find.zipfile.R
\name{url.to.find.zipfile}
\alias{url.to.find.zipfile}
\title{Get URL(s) for FTP site folder(s) with ACS 5-year summary file data}
\usage{
url.to.find.zipfile(mystates, end.year = "2012", url.prefix)
}
\arguments{
\item{mystates}{Character vector of one or more states/DC/PR, as 2-character state abbreviations. Default is all states/DC/PR.}

\item{end.year}{Optional end year for 5-year summary file, as character, defaulting to "2012" but ignored if url.prefix is specified}

\item{url.prefix}{Optional character element that defaults to what is returned by \code{\link{get.url.prefix}(end.year)}}
}
\value{
Returns character vector that is URL(s) such as "ftp://ftp.census.gov/acs2012_5yr/summaryfile"
}
\description{
Returns URL(s) of folders (on Census Bureau FTP site) with zip file(s) based on end year.
}
\details{
The zip files look like this for example: "20135dc0001000.zip" \cr

  The 2009-2013 summary file by state-seqfile combo is in folders that look like this: \cr

  "ftp://ftp.census.gov/acs2013_5yr/summaryfile/2009-2013_ACSSF_By_State_By_Sequence_Table_Subset/DistrictOfColumbia/Tracts_Block_Groups_Only" \cr

  The 2008-2012 summary file by state-seqfile combo is in folders that look like this: \cr


  The 2007-2011 summary file by state-seqfile combo is in folders that look like this:\cr

    "ftp://ftp.census.gov/acs2011_5yr/summaryfile/2007-2011_ACSSF_By_State_By_Sequence_Table_Subset/DistrictOfColumbia/Tracts_Block_Groups_Only/" \cr
  "http://www2.census.gov/acs2011_5yr/summaryfile/2007-2011_ACSSF_By_State_By_Sequence_Table_Subset/DistrictOfColumbia/Tracts_Block_Groups_Only/" \cr
  URL must be the ftp site, not the http version.
  But 2010-2014 was on http only, not ftp, as of mid Dec 3 2015 release day.
}
\seealso{
\code{\link{get.acs}}, \code{\link{url.to.find.zipfile}}, \code{\link{download.geo}}
}
