% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/which.seqfiles.R
\name{which.seqfiles}
\alias{which.seqfiles}
\title{Find Which Sequence Files Contain Given ACS Table(s)}
\usage{
which.seqfiles(tables, lookup.acs, end.year = "2012")
}
\arguments{
\item{tables}{character vector, required. Defines which ACS table(s) to check, such as 'B01001'}

\item{lookup.acs}{data.frame, optional (if not provided then it is downloaded from Census).
Specifies what variables are in which tables and which tables are in which sequence files on the FTP site.}

\item{end.year}{Character element, optional, "2012" by default. Defines end year for 5-year dataset.
Valid years are 2009:2014 as of Dec 2015 - Nov 2016, with more to be added over time.
Ignored if lookup.acs is specified, however. If they imply different years, the function stops with an error message.}
}
\value{
Returns a vector of one or more numbers stored as characters, each defining one sequence file, such as "0001".
}
\description{
The US Census Bureau provides 5-year summary file data from the American Community Survey in sequence files on their FTP site.
  This function reports which sequence files contain the specified tables. Used by \code{\link{get.acs}}
}
\seealso{
\code{\link{get.acs}} and \code{\link[acs]{acs.lookup}} from the \pkg{acs} package, which does something related but is more flexible & robust. Also see \code{\link{get.acs}} which uses this.
}

