% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/zipfile.R
\name{zipfile}
\alias{zipfile}
\title{Get name(s) of zip file(s) for ACS 5-year summary file data}
\usage{
zipfile(mystates, seqfilenum, zipfile.prefix, end.year = "2012")
}
\arguments{
\item{mystates}{Required vector of 2-character state abbreviation(s)}

\item{seqfilenum}{Required single sequence file number used by ACS 5-year summary file}

\item{zipfile.prefix}{Optional character element, defaults to value looked up based on end.year.}

\item{end.year}{Optional end year for 5-year summary file, as character, defaulting to "2012"}
}
\value{
Returns character element that is name of zip file such as "20115dc0113000.zip"
}
\description{
Returns name(s) of zip file(s) based on state(s), a sequence file number, a prefix, and end year.
}
\seealso{
\code{\link{get.acs}}
}

